feat: custom brand loading app based on storefront logo. Set storefront on stores context for accessing the logo.
