+++ js/.tool-versions
@@ -1,1 +1,1 @@
-nodejs 14.17.1
+nodejs 16.10.0
