+++ js/package.json
@@ -8,11 +8,11 @@
   "keywords": [],
   "license": "Apache-2.0",
   "engines": {
-    "node": "~14.17"
+    "node": "16.10.0"
   },
   "scripts": {
     "bootstrap": "lerna link && lerna bootstrap",
-    "build": "lerna run build --scope @oyster/common --scope web",
+    "build": "lerna run build",
     "start:prod": "cd packages/web && yarn run start:prod",
     "start": "cross-env CI=true lerna run start --scope @oyster/common --stream --parallel --scope web",
     "lint": "prettier -c 'packages/*/{src,test}/**/*.ts' && npm run lint:eslint",
