+++ js/packages/cli/package.json
@@ -58,6 +58,7 @@
     "@babel/preset-env": "^7.15.6",
     "@babel/preset-typescript": "^7.15.0",
     "@types/jest": "^27.0.1",
+    "@types/offscreencanvas": "^2019.6.4",
     "jest": "^27.2.0",
     "pkg": "^5.3.1",
     "typescript": "^4.3.5"
