+++ js/packages/common/src/contexts/store.tsx
@@ -9,6 +9,56 @@ import React, {
 import { getStoreID, setProgramIds, StringPublicKey } from '../utils';
 import { useQuerySearch } from '../hooks';
 
+export interface StorefrontMeta {
+  title: string;
+  description: string;
+  favicon: string;
+}
+
+export interface StorefrontTheme {
+  logo?: string;
+  stylesheet: string;
+}
+
+export interface Storefront {
+  pubkey: string;
+  subdomain: string;
+  meta: StorefrontMeta;
+  theme: StorefrontTheme;
+}
+
+export interface StorefrontConfig {
+  storefront: Storefront | void;
+}
+
+export interface ArweaveTag {
+  name: string;
+  value: string;
+}
+
+export interface ArweaveTransaction {
+  id: string;
+  tags: ArweaveTag[];
+}
+
+type ArweaveNode = ArweaveTransaction;
+
+export interface ArweaveEdge {
+  node: ArweaveNode;
+}
+
+export interface ArweaveConnection {
+  edges: ArweaveEdge[];
+}
+
+export interface ArweaveQueries {
+  transactions: ArweaveConnection;
+}
+
+export interface ArweaveQueryResponse {
+  data: ArweaveQueries;
+}
+
 interface StoreConfig {
   // Store Address
   storeAddress?: StringPublicKey;
@@ -18,14 +68,19 @@ interface StoreConfig {
   isReady: boolean;
   // recalculate store address for specified owner address
   setStoreForOwner: (ownerAddress?: string) => Promise<string | undefined>;
+
+  ownerAddress?: StringPublicKey;
+
+  storefront: Storefront;
 }
 
 export const StoreContext = createContext<StoreConfig>(null!);
 
 export const StoreProvider: FC<{
-  ownerAddress?: string;
+  storefront: Storefront;
   storeAddress?: string;
-}> = ({ children, ownerAddress, storeAddress }) => {
+}> = ({ children, storefront, storeAddress }) => {
+  const ownerAddress = storefront.pubkey
   const searchParams = useQuerySearch();
   const ownerAddressFromQuery = searchParams.get('store');
 
@@ -62,7 +117,7 @@ export const StoreProvider: FC<{
   }, [initOwnerAddress]);
 
   return (
-    <StoreContext.Provider value={{ ...store, setStoreForOwner, isConfigured }}>
+    <StoreContext.Provider value={{ ...store, setStoreForOwner, isConfigured, ownerAddress, storefront }}>
       {children}
     </StoreContext.Provider>
   );
