+++ js/packages/web/src/App.tsx
@@ -1,12 +1,13 @@
 import React from 'react';
+import { Storefront } from '@oyster/common';
 import { Routes } from './routes';
 
 interface AppProps {
-  storeId: string;
+  storefront: Storefront;
 }
 
-function App({ storeId }: AppProps) {
-  return <Routes storeId={storeId} />;
+function App({ storefront }: AppProps) {
+  return <Routes storefront={storefront} />;
 }
 
 export default App;
