+++ js/packages/web/src/actions/getStorefront.ts
@@ -2,7 +2,7 @@ import {
   ArweaveTransaction,
   Storefront,
   ArweaveQueryResponse,
-} from './../models/storefront';
+} from '@oyster/common';
 import { maybeCDN } from '../utils/cdn';
 
 const ARWEAVE_URL = process.env.NEXT_PUBLIC_ARWEAVE_URL;
@@ -61,12 +61,18 @@ export const getStorefront = async (
         stylesheet: maybeCDN(`${ARWEAVE_URL}/${transaction.id}`),
       },
       meta: {
-        favicon: maybeCDN(values['holaplex:metadata:favicon:url'] || "/favicon-16x16.png"),
-        title: values['holaplex:metadata:page:title'] || `Holaplex - ${subdomain} | NFT Marketplace`,
-        description: values['holaplex:metadata:page:description'] || 'A NFT marketplace generated by Holaplex',
-      }
-    }
+        favicon: maybeCDN(
+          values['holaplex:metadata:favicon:url'] || '/favicon-16x16.png',
+        ),
+        title:
+          values['holaplex:metadata:page:title'] ||
+          `Holaplex - ${subdomain} | NFT Marketplace`,
+        description:
+          values['holaplex:metadata:page:description'] ||
+          'A NFT marketplace generated by Holaplex',
+      },
+    };
   } catch {
-    return null
+    return null;
   }
-}
+};
