+++ js/packages/web/src/styles/app.less
@@ -58,6 +58,30 @@ body {
   }
 }
 
+.app {
+  &-auctions-list {
+    &--loading {
+      text-align: center;
+      padding: 30px 0 0 0;
+    }
+  }
+  &--loading {
+    &-logo {
+      width: 42px;
+    }
+    position: fixed;
+    top: 0;
+    left: 0;
+    bottom: 0;
+    right: 0;
+    text-align: center;
+    display: flex;
+    align-content: center;
+    justify-content: center;
+    align-items: center;
+  }
+}
+
 .no-events {
   pointer-events: none;
 }
